Move `getDistTag()` function into its own file
Because the `builder-cache.ts` file has side effect when imported,
which we don't want to happen during the build script.
